Show rebuild status in the IDE.
Summary:
Problem: There was no user-visible change when we rebuilt the analysis directory after detecting a rebase.

The client now sends a `ShowStatus` request to the server when it starts rebuilding. This shows a `Rebuilding` status message in the IDE.

When the rebuild is finished, it sends a "Done rebuilding" message.

Reviewed By: sinancepel

Differential Revision: D18846642

fbshipit-source-id: 9ac7d2b6181fd7433ee7fdd5fac9f5982789c863