Don't store locations in deobfuscation
Summary:
Later in this stack I'm going to make Node.equal location_sensitive.  The best way to avoid comparing locations when you don't want to is to not store them, so by converting names into references and storing everything else directly, we sidestep the problem entirely.
The only caveat I had to make was an assumption that all callable types would be simple names.  This doesn't work for higher order functions or for local lambdas, but I'm not sure those really count as constants anyway.  In any case making that change didn't cause any tests to fail so I'm going to assume that this is good enough.

Reviewed By: grievejia

Differential Revision: D18905702

fbshipit-source-id: 9fce5829e30cb696051e4b18edbf1b3e03b3e20d