Store location information in class names
Summary:
LibCST is requesting type information on class names.

 ---

Original explanation on the difference between expressions and references: D14487803

I felt it would be most appropriate to create reference nodes here, because we still care about the simplicity that references lend us in the backend logic (ie., chains of strings only, no calls or other sub-expressions supported).

However, the original quote that "These are not themselves part of the AST, but they are metadata we attach to parts of our backend that reference something else." no longer entirely applies - the name of a class is part of the AST, but other references in the pyre backend (ie., in metadata) are actual references to other concrete objects.

Reviewed By: sinancepel

Differential Revision: D18910110

fbshipit-source-id: a70eaa2c8dfa2ac65ab741f90f7be5f2264470e1