Fix Vim-Watchman bug by changing the buck query logic.
Summary:
Problem: When we used Vim in a project that defined `srcs` in the `python_binary(name='foo')` target itself, instead of having a separate `python_library`, then Buck returned the owner target as `foo-library` instead of `foo`. This meant that we could not detect the file when it was added back after being deleted (thanks to Vim's swap file logic).

We now check for the new `-library` suffix as well.

Logged the buck query command.

Reviewed By: grievejia

Differential Revision: D18939820

fbshipit-source-id: 47441f0fe6326a2d111fd8aea4dfb88d3b484f87