Create nonexistent directory for file updates if they do not already exist
Summary:
Turns out that the `tee` command would fail if we are teeing into a file that lives in a nonexistent directory. This would break all file updates as we do not check parent directory existence for them.

Let's fix this by running `mkdir -p` on the parent directory for all updated files, if there exist any.

Reviewed By: sinancepel

Differential Revision: D18948063

fbshipit-source-id: fcdfabb8bd61067b6801162b4bcd401a1f0f48b3