Fix bug in the logic for testing if a new file was tracked.
Summary:
Problem: To check whether the owner target of a new file was tracked by the project, I did a string comparison against the target. Bad idea. It meant that it wouldn't match variations on the target, such as `//tools/pyre/...` vs `//tools/pyre/client/commands:commands`

Use the Buck `intersection` operator.

This also fixes the Vim-Watchman bug in general. However, there is no good way to handle the edge case when `srcs` are defined on `python_binary`. Using an intersection with the project-level target returns empty results. So, I'll go ahead with this and recommend that the one person who reported having a `python_binary` with `srcs` should split it up into `python_library`.

Reviewed By: sinancepel

Differential Revision: D19000677

fbshipit-source-id: 516fb73a2f510d7ed40f8c552786581b6be59b11