Add support for import type query lookups
Summary:
Support type reveals on imports. These are a little special because they're not local annotation lookups, and so go in the special-cased override lookup function to handle qualification.

Note - we currently don't support types for modules (e.g. `reveal_type(typing)`) - they are always unknown. Made a wishlist backlog task to type these as `typing.ModuleType` in the future but didn't think that was urgent now. T59139164

Reviewed By: dkgi

Differential Revision: D19053395

fbshipit-source-id: b69e15d01427e46521b1fde8c74a4f968d868737