Also track changes in dependencies of tracked targets.
Summary:
Problem: We didn't consider a new file to be tracked if it was part of a dependency of our main target.

Use `deps()` to also check the new file against dependencies.

This fixes the edge case in the Vim-Watchman bug where the `srcs` were defined on `python_binary` instead of `python_library`. Earlier, this would make `buck query` return the owner target as `foo-library`, which would not be considered as part of `foo`.

Reviewed By: sinancepel

Differential Revision: D19129538

fbshipit-source-id: 58e566edfa2043565c942cbe67c3b0e22f444e60