simulate calls backing comparison operators
Summary:
We were previously not analyzing the calls corresponding to comparison operators at all. (a == b morally gets transformed to a.__eq__(b) for the purposes of type checking, with fallbacks in some cases). Add some coverage here by re-using the conversions from type checking.

This isn't a problem for (1 + 2), etc., because we do the conversion in the parser for those - not sure why we do that, but it's the state of the universe...

Reviewed By: dkgi

Differential Revision: D19128743

fbshipit-source-id: 3d4768dd9f7ec60a403b77e95db006ac1f51592a