include static functions invoked with partial() in taint analysis call graph
Summary:
We want to catch flows going through `partial` in the taint analysis. This lays the groundwork by
enriching the callgraph to include the actual functions when partial is invoked.

The end goal is to catch flows from the partially applied parameters, i.e., if we have def foo(flows_to_sink, benign), and call `partial(foo, tainted_argument)`, we want to mark `partial()` as a point where an issue occurs.

Reviewed By: fahndrich

Differential Revision: D18991850

fbshipit-source-id: a8bfaa27d79132b46efbaab31d7732741b717d3d