add type breadcrumbs for obscure models in forward analysis
Summary: We generally add type based breadcrumbs when writing the models for a callable we just analyzed. This works for most cases, but fails for obscure models - when we get an obscure model, we previously never had a chance to create a type based breadcrumb. Follow what we do in the backwards analysis and compute type based breadcrumbs on calls for obscure models.

Reviewed By: fahndrich

Differential Revision: D19126946

fbshipit-source-id: 58a09ca2eccaa4b8ee6deed000c3a5da3e9b27a3