Move setup.py into its own file from a shell script for the OSS build
Summary:
Context: I'm looking to make some changes into setup.py to also package taint models, so this felt like an appropriate time to tame the beast.

This diff moves setup.py from being generated in a shell script into a regular file that we copy to the build directory when running the build commands. As part of the fallout, some environment variables are inlined, and the one remaining environment variable, which is the PACKAGE_VERSION one, is read through os.environ instead (build-pypi-package.sh has to export the variable as a result).

We're giving up the ability to add typeshed as a runtime dependency to Pyre as a result of this (i.e. this forces us to bundle typeshed when releasing Pyre). The reason is that I felt the complexity of supporting that wasn't worth it, and we're always bundling typeshed anyway.

Reviewed By: MaggieMoss

Differential Revision: D19153293

fbshipit-source-id: 8cd6e3cdbd18d40a45271ed52c00e7c9fe5adf26