Include the dependency table in the saved state
Summary:
We were under the impression that the saved state creation functionality in hack_parallel included the deptable in its saved state.  Unfortunately this is not the case.

The most straightforward and safe way to remedy this is to tar together the dependency table saved state and the hash table saved state, and untar them on the other side.  Otherwise we would need to alter the entire pipeline to manage a directory rather than a file.

Reviewed By: grievejia

Differential Revision: D19203684

fbshipit-source-id: 55388f1c1894f163df01973fa9ad36cac230ecdd