Properly grab function definitions from Pyre
Summary:
I discovered this issue while working on another unrelated diff.

It looks like the intention of `Callable.get_definition` is to extract the body of a function with a given name. The "officially blessed" API for this purpose from the Pyre side is `GlobalResolution.function_definitions` (introduced in D18720713), which maps from function names to their bodies.

Existing implementation of `Callable.get_definition`, however, did something creative: when the function is actually a method defined in a class, it attempts to grab the function body by first looking up the body of the class and then does a linear scan of the class body trying to find a function whose name matches. This implementation is undesirable for two reasons:

- On the Pyre side, both class bodies and the methods bodies defined in them are not indexed. This means that the "get the body of the class with a given name" operation and the "get the body of the method within a class with a given name" operation are both very inefficient as they require linear scan of their containing source/class. In fact, the operations were so inefficient that a cache was specifically introduced just to speed things up.
- All the work around getting class definitions is completely useless. In Pyre, methods are treated in the same way as functions, and their bodies can be directly obtained through `GlobalResolution.function_definitions` -- there's no point to go through class definitions for them.

This diff tries to make the implementation of  `Callable.get_definition` less awful by building the method lookup logic directly on top of function definition lookup.

Reviewed By: fahndrich

Differential Revision: D19221810

fbshipit-source-id: 1bc653c13fba48d7f9c8f6e2fd4638ca6b6973d3