Stop storing module names in `Location.t`
Summary:
This diff is actually less scarier than it looks. All I want to do here is to take the `path` field out of `Location.t`.

Context
======

Storing the `path` field in `Location.t` is problematic for the following reasons:

- It's wasteful in terms of sharedmem usage.
Since `Location.t` is included in all `Node.t`, when serializing AST to sharedmem we need to re-serialize the same path over and over again each time a `Node` is encountered in the AST. To make matter worse, we also dump our AST multiple times (i.e. `AstEnvironment.RawSources`, `AstEnvironment.Sources`, and `UnannotatedGlobalEnvironment.FunctionDefinitions`) to the sharedmem.

- There's also little value in keeping the `path` field in `Location.t`. Modules/paths are actually not used at all until error reporting. Whenever errors need to be reported, module names / path names could be obtained from sources other than `Location.t`.

Note that the same observations has been made previously in D8743535, but the solution there was to hash all the paths, which does not play well with dependency tracking (and therefore gets reverted in D16648821). I think we are now in a better position of implementing a more aggressive solution.

Implementation
===========

The basic idea is that `Location.t` should only contain the start and stop position. If an additional `path` field is useful, then let's construct a different data structure for it. Changes to `location.ml` can be summarized in the following table:

| Type      | Changes |
| ----------- | ----------- |
| `Location.t`      | No longer alias to `Location.Reference.t`: the `path` field is dropped.       |
| `Location.Reference.t`   | Renamed to `Location.WithModule.t`        |
| `Location.Instantiated.t`   | Renamed to `Location.WithPath.t`        |

The rest of this diff is all about dealing with the consequences of these changes.

Reviewed By: sinancepel

Differential Revision: D19234094

fbshipit-source-id: 2adad9ebbc71c40267a0f7819917ec2fcbe408e2