Start a configuration monitor for each server.
Summary:
Problem: D18892575 introduced one project-level configuration monitor. However, this means that:

+ When the local configuration for a second server is changed, it gets ignored and no server is stopped.

+ When a second server is started, it is unable to acquire the monitor lock and create a new monitor. This means that the old monitor will keep running and we won't be able to upgrade the monitor until we stop the old monitor manually.

+ We are unable to stop the monitor in `Stop` or `Kill` because there may be other servers that depend on the monitor.

Solution: Revert back to the old behavior of one configuration monitor per server. This will allow us to stop the corresponding server when a local configuration is changed, start new monitors, and stop the monitor when the server is stopped.

Reviewed By: shannonzhu

Differential Revision: D19238855

fbshipit-source-id: bb676305ad834eb8ce23d2a9ececfa5c6eb4257a