stop flows for `x = f(x)` in backward analysis
Summary:
In backwards analysis, we weren't clearing the taint of `y` in the assignment of `y = f(x)`. This is fine in most cases, but consider the following:

```
def foo(x):
  x = untainted(x)
  sink(x)
```

In this assignment, because `untainted(x)` assigns no taint (and is not obscure), `x` should now be free of taint. However, because we previously didn't clear the taint of `x`, we would mark a flow from the `x` parameter to a sink.

Solve this problem by deleting the taint of `x` after reading the taint and before analyzing the `untainted(x)` expression.

This is not an issue in forward analysis, because we unconditionally store the taint we get from `untainted(x)` to the access path corresponding to `x` .

Reviewed By: fahndrich

Differential Revision: D19239658

fbshipit-source-id: bd0e6ccba4019d7e671d1a2df3ac8b82ec50e693