Consistently treat __new__ as static method
Summary:
We need to make sure we aren't erroring in calls like
`temp = self.__class__.__new__(self.__class__)` (example from fb_record).

Our previous system stripped the `cls` parameter and `self` parameter from classmethods and instance methods when returning them as attributes, but also stripped the first parameter out of `__new__` which is a static method.

This works fine for say, `__init__` but breaks when `__new__` is explicitly called. In those cases, you actually are expected to pass in an explicit `cls` parameter but we would error that `__new__` expects 0 parameters and it was given 1.

This diff does a couple things:
1. ensures the `cls` parameter is not stripped from `__new__` when returning it as an attribute
2. adds on the expected parameter type of `cls` to be `Type[parent_type]` when returning it as an attribute
3. strips the `cls` parameter from `__new__` when it's looked up as an implicit constructor, ie., when `Foo()` is called and `class Foo` defines a `__new__` rather than an `__init__`. This is because implicit calls to `__new__` automatically apply the `cls` parameter, and this was a cleaner solution for now than trying to filter out this case in forward resolve callable and create an argument to feed into cases where the returned implicit constructor is a `__new__`.

Reviewed By: dkgi

Differential Revision: D19237322

fbshipit-source-id: dc9185dbb73ecebed23ba6d51923cfb2e3196c44