unbreak `pyre init`
Summary:
Turns out that D18802372 broke `pyre init` because we still unconditionally assume that all commands have a analysis directory and configuration. This is no longer true for Initialize, which, after the refactor, inherits directly from CommandParser instead of Command.

Pyre didn't catch this because the annotation of `Optional[Command]` for command is wrong - it is actually a CommandParser, and we have a bit of magic dynamism left still.

Fix the annotation + work around the issue.

Reviewed By: grievejia

Differential Revision: D19316639

fbshipit-source-id: 733de01c78a782ea91a492622e269e9395fc17a6