Collect type variables from define using uncorrected parse
Summary:
In order to error on type variables not in the parameters of a function (or an outer class/function), we collected the type variables from a parse of the function definition that implicitly ran the invalid type parameters check. The invalid type parameters check replaces invalid parameters with Anys.  If those parameters were type variables, or contained type variables, this meant that they were removed from the parameters, meaning they were never noted as being in scope.  This meant that the subsequent type variable in scope check (done during a separate parse) would spuriously determine that those variables were not in scope.  Due to the sequencing of checks/correction, this error clobbered the invalid parameter error, leading to *only* getting a manifestly incorrect "variable not in parameters" error.

The fix is to do the variable collection off of a parse that does not run the invalid type parameters check

(detail: under the hood, the invalid type parameters check is unconditionally run on alias "values", meaning that even if the variable itself contains an invalid type it will consistently be parsed the same way and thus match when doing the downstream check)

Reviewed By: sinancepel

Differential Revision: D19314813

fbshipit-source-id: fb2915643dfd733fc2432634ab349d7ef1493b14