Drop nested function&class bodies in nesting defines
Summary:
The motivation here is to save shared memory space, as we've seen an increase in the number of `Sharedmem.Heap_full` exception thrown on prod recently.

According to `pyre --debug check`, the largest table in shared memory is the table that stores all defines for type checking: it's roughly 5 times as large as the table that stores all the ASTs, which is weird.

One potential reason why this is the case is that nested function/class bodies are stored multiple times in the define table: consider the following file `foo.py`:
```
def bar() -> None:
  def baz() -> None:
    ...
  class Qux:
    ...
  ...
```

We used to preserve the full body of `foo.bar.baz` and `foo.bar.Qux` in the body of `foo.bar`, even though they are completely unused in the type checking phase. Let's fix that by stripping away all those nested bodies before putting `foo.bar` into shared memory.

Reviewed By: sinancepel

Differential Revision: D19312681

fbshipit-source-id: 24e1d7df6dfd2e4a1cb2902c7ed3c7ec0e8bb08b