support all recognized decorators
Summary:
The `Define.is_decorator` signature is deprecated, and what we really need to do is use the Recognized.property_decorators set to determine whether a decorator matches.

I initially tried to re-use Pyre's understanding of properties instead of re-rolling the logic myself. This didn't work out, as the properties you get out of the Attribute module don't contain a reference to the implementing signature, which we need to verify the names of parameters.

Instead, I removed the is_property API from statements (since it's not used anywhere else) and inlined the correct implementation to model.ml.

Reviewed By: dkgi

Differential Revision: D19304977

fbshipit-source-id: b9ace7fa25a0caf25d883772feb20721d8c90bec