start executing unused tests and kill unused toplevel values in tests
Summary:
Note: This is part 1 of adding mli's to all tests so that we don't have un-executed test functions. To make it easier to review, I'm breaking it up into two parts - one where we fix all violations, and a second where we add the mli's.

I was able to preserve most tests, but one test was egregiously wrong (almost all cases were failing), so I removed that one.

Reviewed By: grievejia, mrkmndz

Differential Revision: D19332588

fbshipit-source-id: 7e791f46129e52fb7c60baa151ddaa14425540eb