Only count a function as untyped when both arguments and return are unannotated
Summary:
According to [PEP484](https://www.python.org/dev/peps/pep-0484/#the-meaning-of-annotations):

> Any function without annotations should be treated as having the most general type possible, or ignored, by any type checker.

Also,

> For a checked function, the default annotation for arguments and for the return type is Any.

I think this implies that if none of a function's arguments and return is annotated, we should not report any errors in them. The current behavior of `mypy` confirms it. However, this is not what Pyre does today, which controls the reporting behavior based on return annotations only. Let's fix that.

The fix in Pyre is straightforward. The unexpected impact was on the Pysa side, where more type check errors are ungated in the integration test suites, which need to be patched as well.

Reviewed By: dkgi

Differential Revision: D19330883

fbshipit-source-id: 8e0a3707ba864f9c7e1ddfc628ef56c113a8f3c4