Actually apply decorators to nested functions
Summary:
It turns out we were completely ignoring decorators on nested functions :/.

I think this attributable to the confusingness of the various define -> Type functions floating around that all have similar names.  For now I'm just fixing this problem, and will take a second look in a follow-up to see how to make this less confusing.

Reviewed By: sinancepel

Differential Revision: D19349700

fbshipit-source-id: 84a9a14b19e7088604a2e7f231eb4aa41e4b3835