turn FunctionTainter into a component instead of a base class
Summary:
We need to turn FunctionTainter into a component to make the GetRESTApiSources model generator functional - since it already is a view generator, we'd otherwise have to call __init__ on each superclass. (Independently, I find multiple inheritance quite magical).

It was relatively straightforward to do so, and you can already see a win - the interface of creating the GetRESTAPISources generator is cleaner now, as evidenced in the `get_filtered_sources.py` file, since we don't need to repeat ourselves and can control the interface more explicitly.

Reviewed By: gbleaney, C0deBr8kr

Differential Revision: D19322829

fbshipit-source-id: dcb4307776392442cec6cf3f4d8ac4f506c610ae