Supporting capture groups
Summary:
from the discussion on D16663193:
> The reason we're creating a variadic out of two non-homogenous variable inputs to Tensor is because we don't want to deal with the syntax mess introduced by having to interpret a list of variable inputs into multiple separate variables and lists of variables - unary(ies) and variadic(s). Doing that would require some kind of sub-list syntax when instantiating the type and its parameters, and it would also allow the system to handle classes that are generic over more than one variadic, which is not supported now (and is why there's a concept of a single Middle in the concat structure)...
Concat just happens to be useful for expressing classes that are generic over unaries and one variadic without having to introduce even more syntax as described above.

The time has finally come to "deal with the syntax mess" :).

The syntax implemented here is the following:

(T1, T2, ... Tn are all TypeVars, and T1s, T2s, ... Tns are all ListVariadics)

| #| Declaration site |  Instantiation Site | Remarks
|--| -- | --| --|
|1| `class C(Generic[T1, T2, T3])` | `C[int, str, bool]`| This is the existing syntax for declaring generic classes
|2| `class C(Generic[T1s])`| `C[int, str, bool]` or `C[Ts]` or `C[Map[List, Ts]]` etc.| This was supported by the old implementation, and is how we need to continue to look, in order to maintain similarity to the built-in variadic classes (`Tuple`, `Union`, etc.).
|3|`class C(Generic[T1,T1s])`| `C[int, [str, bool, float]]` or `C[int, Ts]` or `C[int, Map[List, Ts]]` etc.| This replaces `Generic[Concatenate[T, Ts]]` for the tensor case.
|4|`class C(Generic[T1, T1s, T2s])`| `C[int, [str], [bool, float]]`| This is brand new functionality, as the previous implementation couldn't allow you to be generic on two different variadics at once.

Rationale:
* The existing Concatenate solution is even more confusing than I anticipated even when making the initial compromise.
* Relying on Concatenate in declaration means that you have to rely on Concatenate in instantiation whenever you want to propagate the variable.  This exercises logic that is intractable to implement in the general case (Concatenation against Concatenation), for something that should be very easy (just a dtype and a shape).  This makes common cases currently fail (see footnote 1 in classHierarchyTest).
* The existing system would be incredibly hard to extend to support ParamSpec Generics, while the new system is almost easier to do with them than without them.

Downside:
One aspect of this I'm less than pleased with is the OCaml representation of #1 vs #2.  The core issue I have is that you can't know syntactically whether `C[int, str, bool]` is for a class declared like #1 or like #2.  In theory we could require that you instantiate #2 as `C[[int, str, bool]]`, but this looks kind of weird, and worse, doesn't match with how `Tuple`s and `Union`s are instantiated (which we won't want to change).

That means that we have to convert from `[Single a, Single b, Single c]` to `[Group [a, b, c]]` on demand when attempting to do a zip (see ClassHierarchy.Variable.zip_on_parameters and zip_on_two_parameter_lists).  This unfortunately runs afoul of one of my classic OCaml rules: don't allow for semantically equivalent things to have multiple representations.  In theory I could fix this by invalidating one of those options (which would have to be the second in order to preserve normal behavior in the unary cases).  Such a type would have to look something like this
```
type one =
| OSingle of Type.t
| OAny
| OConcatenation of OrderedTypes.Concatenation.t

type multiple_member =
| Single of Type.t
| Group of OrderedTypes.t

type parameters =
| Empty
| One of one
| Multiple of { head: multiple_member; middle: multiple_member; tail: multiple_member list}
```

Even for me who is prone to being dogmatic about that rule, this seems like overkill.  For now, this seems to be a benign issue that can be handled outside of the OCaml type system in actual application logic.

Reviewed By: pradeep90

Differential Revision: D19296079

fbshipit-source-id: c17e8effdc148242577a9821a81896b44f4387be