Correctly supporting meta-type creation for some special cased "classes"
Summary:
Unfortunately in Python type expressions are present in both annotations, and in arbitrary code.

To handle this second case we try to guarantee that resolve(TypeExpression) = typing.Type[Type.create(TypeExpression)].

This becomes complicated with generics, and even more complicated with the special built-in generic classes.

I have a follow-up diff to support user-defined-variadic meta type handling, but we can split out this work supporting some "natural" variadics.

Reviewed By: pradeep90

Differential Revision: D19415297

fbshipit-source-id: e470f56ad314be69f12696d2d9debe4d7107b026