Support explicit generics as literals
Summary:
We currently support unannotated globals of the following form
```
x = NonGenericClass(some, complex, arguments)
```
This is because we can tell that it will return a `NonGenericClass` no matter what the arguments are.

We correctly do not support
```
x = GenericClass(some, complex, arguments)
```
because we need more information than we have at the global registration phase.

This diff introduces the ability to register
```
x = GenericClass[int, str](some, complex, arguments)
```
as `GenericClass[int, str]`,  as a specialized generic constructor is basically just a concrete constructor.

Reviewed By: sinancepel

Differential Revision: D19386183

fbshipit-source-id: 458b181760a67c9e3ac8db43d4baf9e533b49239