Introduce APIs for parallel scheduling policy
Summary:
We'll need to go through a fair amount of background before one can make sense out of my proposal here.

Context
======

The problem we are trying to address here is to figure out a reasonable way to schedule the list of workloads passed into `Scheduler.map_reduce` (or `Scheduler.iter`). Past experience has shown that the scheduling algorithm has critical impact on the performance profile of both `pyre` and `pysa` (examples where a simple decision change lead to hundreds of seconds of perf win: D19385381, D13330658).

Here's a little bit of background on how the underlying `MultiWorker` module in `hack_parallel` is coded up. Supposed that we have `M `tasks and `N` workers. Conceptually, the way it works is that we take the list of tasks, break it apart into smaller chunks (aka. `bucket` in the code), and put all the chunks into a queue. Then we repeatedly poll from a pool of workers -- if all workers are busy, we block and wait for one that becomes non-busy; otherwise, we pull a task out of the queue, assign it to a non-busy worker, and poll for the next non-busy worker.

As you can see from the description above, there's really not a lot we can fine-tune given this framework -- picking a scheduling algorithm, under this context, essentially boils down to deciding how the list of `M` input tasks can be chunked.

Although the underlying `MultiWorker` API makes it possible to implement more advanced chunking algorithm, I think to avoid over-engineering we'll just stick with the simplest and dumbest chunking method possible for now: splitting the `M` jobs into chunks of equal length. Given that the input size `M` can vary across different invocations of `Scheduler.map_reduce`, naturally there are two ways we can do the split:
- The Fixed-Bucket-Size policy keeps the size of each chunk (henceforth denoted as `B`) constant, and let the number of chunks vary as `M / B`.
- The Fixed-Bucket-Count policy keeps the number of chunks (henceforth denoted as `C`) constant, and let the size of each chunk vary as `M / C`. Note that if this kind of policy is used, it usually makes more sense to let `C` scale with the number of workers -- hence in the code you won't find  `C` directly because it is always derived by multiplying the worker count and a multiplier.

Choosing Policies
=============

I've thought a lot about how we should choose between the Fixed-Bucket-Size policy and the Fixed-Bucket-Count policy. My current understanding is that there's no clear winner and it's always about tradeoffs:
- The Fixed-Bucket-Size policy tends to have more predictable **latency**
  + It's easier to estimate how long it takes for a worker to complete a given chunk of work if the chunk size is known.
- The Fixed-Bucket-Count policy tends to have more predictable **throughput**
  + It's easier to estimate the amount of IPC communication overhead if we know how many chunks need to be processed in total.

Following this line of thoughts, the first principle we should follow seems to be that we should do Fixed-Bucket-Size whenever progress needs to be shown to the user (because humans are more latency sensitive), and do Fixed-Bucket-Count otherwise for better overhead control. Note that I'm simplifying the issues a lot -- there are more perf/memory concerns regarding the worker GC if we are willing to dig into more details. But I want to get the big picture right first, before we can dive into micro-optimizations.

Fallback Behaviors
==============

Unfortunately, I can't conclude the discussion of scheduling policy at this point, as we do have some devils that hide in the edge cases.

The simple Fixed-Bucket-Size and Fixed-Bucket-Count policies described above work great when the number of task `M` is sufficiently large. But when the number becomes small, we run into troubles fairly quickly:
- The Fixed-Bucket-Size policy is prone to the problem of *starvation*: when the number of work `M` is not sufficient enough to fill `N` buckets (with `N` being the number of workers) of size `B`, some worker is going to be left with little to none work to do, while other workers will keep getting `B` chunks of works.
- The Fixed-Bucket-Count policy is prone to the problem of *excessive overhead*: when the number of work `M` is not sufficient to the extent that each bucket only contains very few work to do, the IPC/synchronization overhead between master-worker starts to dominate the wall clock. I've seen in practice cases where for a small amount of work, doing everything in master takes milliseconds while sending the same workloads to workers would take tens of seconds.

To mitigate these two issues, I propose two simple adaptations to the two policies mentioned before:
- First, we introduce the notion of *fallback behavior*, which kicks in when we find that following the original Fixed-Bucket-Size or Fixed-Bucket-Count policy may lead to issues due to insufficient input size.
  + For Fixed-Bucket-Size policy, if we detect that the number of buckets would be smaller than a threshold `C`, we fall back to a Fixed-Bucket-Count policy of count `C` instead to avoid starvation. The other way of looking at this is that we always cap the number of buckets to be at least `C`.
  + For Fixed-Bucket-Count policy, if we detect that the bucket size would be smaller than a threshold `B`, we fall back to a Fixed-Bucket-Size policy of size `B` to avoid excessive overhead. The other way of looking at this is that we always cap the bucket size to be at least `B`.
- Second, we introduce the notion of `sequential_cutoff`. This is going to be a parameter of our policy, and it means that if the input size `M` is smaller than `sequential_cutoff`, we unconditionally run everything sequentially in the master process. The exact value of `sequential_cutoff` may vary: if each task takes long time to compute (e.g. Pysa fixpoint iteration), we should use a small number; if each task only takes miliseconds (e.g. most of the environment table update), we should use a larger number. Note that this number should always be larger than 1: if `M` is 1, it is unconditionally better to do it in master than in worker.
  + In the code, `sequential_cutoff` is implicitly computed by multiplying the minimum bucket size `C` and the minimum bucket count `B`.

Overview of This Diff
===============

This diff tries to provide implementations for both Fixed-Bucket-Size-WIth-Fallback policy and the Fixed-Bucket-Count-With-Fallback policy. It also preserves the two policies currently used in production (denoted as "legacy" policies in the code) -- the goal is to keep this particular diff semantic-preserving, and gradually phase out the legacy policies in followup diffs.

See `schduler.mli` first to get an idea of how the `Scheduler` API evolves. My change should be fairly straightforward: the `bucket_size` parameter (which used to switch between our legacy version of Fixed-Bucket-Size and Fixed-Bucket-Count policy) is replaced with a full-blown policy type `Policy.t`.

The type `Policy.t` is kept private to ensure that one can only create them via the provided APIs. Under the hood it is nothing but a function that computes the bucket count.

Rest of the changes are there to cope with the `Scheduler` API updates.

Reviewed By: sinancepel

Differential Revision: D19402492

fbshipit-source-id: a9b627cdf9af083b021fe51e825168dc3abfaa45