Replace legacy scheduling policy for environment update
Summary:
Our legacy policy does not leverage sequential cutoffs and fallbacks. As a result, it makes sub-optimal decisions when the number of keys to update is small.

Let's patch the issue with the new policy. I used my gut-feelings to set the parameters, which suffices for a quick patch for now. I'll definitely follow up with more large-scale incremental tests in the future for more fine tuning.

Reviewed By: MaggieMoss

Differential Revision: D19450146

fbshipit-source-id: ac6e3729732e49d789ec79e83e1fa1126d81f456