Fix pyre --version in nested directories
Summary:
We were only successfully finding the binary version when our current working directory was the project root. :'(

This is another upsetting result of doing the chdir() and then relying on a constant `.pyre_configuration` without any qualification to find the project config. Fixing this first to unblock integration tests, hopefully we can do some more client refactoring after to get rid of this.

Reviewed By: dkgi

Differential Revision: D19452484

fbshipit-source-id: 9d1d19674987df3aad691aecc125fa3da18b6fb2