Replace legacy scheduling policy for type checking and postprocessing
Summary:
Our legacy policy does not leverage sequential cutoffs and fallbacks. As a result, it makes sub-optimal decisions when the number of keys to update is small.

Let's patch the issue with the new policy. I used my gut-feelings to set the parameters, which suffices for a quick patch for now. I'll definitely follow up with more large-scale incremental tests in the future for more fine tuning.

Reviewed By: shannonzhu

Differential Revision: D19450279

fbshipit-source-id: d84de04bcc49322a2f0c4aba1c7c83b6d7a3191e