Suppress undefined attribute error for attribute writes when `__setattr__` is defined to accept `Any`
Summary:
This feature is a user request. See linked task for more details.

I settled on a decision where we would only look for "setattr any" pattern for the given class, not any of its parent. This means that if class `Foo` defines a `__setattr__` that accepts `Any` and class `Bar` is a subclass of `Foo`, then we won't suppress undefined attribute error when you do `bar.non_existent_attribute`. The reason is that if were to take all parent class into accounts, then all undefined attribute errors would be suppressed because of [the `object` is defined to have "setattr any" on `typeshed`](https://github.com/python/typeshed/blob/master/stdlib/2and3/builtins.pyi#L54).

Reviewed By: pradeep90

Differential Revision: D19443279

fbshipit-source-id: 313cc3121afbd3b0c269038c29d886ee4c8aea52