allow `defines` queries to target classes as well
Summary:
THis change extends the `defines` query to be able to operate on a class as well.

Chose to special-case this instead of allowing arbitrary filters so that

a) behavior is consistent and b) we can optimize the implementation of the query later if need be.

Hopefully, this allows us to deprecate the `methods()` query.

Reviewed By: grievejia

Differential Revision: D19433403

fbshipit-source-id: e7ba00b1cbfbed58cfdd68ce2b8e91a8444e35e6