No longer store a reference to the update result chain inside of an <X>Environment.ReadOnly
Summary:
We were storing the previous_update_result via closure inside of hash_to_key_map. This meant that when we were marshalling the job closures (which usually involve environments), we were indirectly storing the complete list of previous dependencies.  For large updates, this was a humongous overhead for spawning jobs, meaning that parallelism was actually costing us.

To fix this, we remove the offending closures, and find a subtly different way to get access to what we need.

Reviewed By: grievejia

Differential Revision: D19509235

fbshipit-source-id: a7e3f47995dbc9f21857153786a4959428b4eaf5